Add to Combinatorial Circuits
Adding a section of an equality circuit (as part
of a more general magnitude comparator) for a couple
of reasons.

First is that it's bothered me for a while that
we show building all the standard circuits without
any indication of how to figure it out. So I've known
for a while that I needed some simple circuit which
would serve as an example.

Second is that I want to use the equality checker
as part of timer triggering, so that set the circuit
I am doing.

I also made a few adjustments to help the text flow
a little smoother.